// src/types/categories-amnesia.ts
// Generated by https://quicktype.io
// endpoint:
// https://cms.aipus.co/api/subcategories?populate=media&filters[category][id][$eq]=6

export interface CategoriesAmnesia {
    data: CategoriesAmnesiaDatum[];
    meta: Meta;
}

export interface CategoriesAmnesiaDatum {
    id:         number;
    attributes: PurpleAttributes;
}

export interface PurpleAttributes {
    id_subcategory: string;
    uuid:           string;
    name:           string;
    description:    string;
    deleted:        boolean;
    created:        string;
    updated:        string;
    createdAt:      string;
    updatedAt:      string;
    publishedAt:    string;
    locale:         string;
    data:           null;
    media:          Media;
}

export interface Media {
    data: MediaDatum[];
}

export interface MediaDatum {
    id:         number;
    attributes: FluffyAttributes;
}

export interface FluffyAttributes {
    name:              string;
    alternativeText:   null;
    caption:           null;
    width:             number;
    height:            number;
    formats:           Formats;
    hash:              string;
    ext:               string;
    mime:              string;
    size:              number;
    url:               string;
    previewUrl:        null;
    provider:          string;
    provider_metadata: null;
    createdAt:         string;
    updatedAt:         string;
}

export interface Formats {
    thumbnail: Thumbnail;
}

export interface Thumbnail {
    name:   string;
    hash:   string;
    ext:    string;
    mime:   string;
    path:   null;
    width:  number;
    height: number;
    size:   number;
    url:    string;
}

export interface Meta {
    pagination: Pagination;
}

export interface Pagination {
    page:      number;
    pageSize:  number;
    pageCount: number;
    total:     number;
}